# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_analyze_swatch_foundations_analyze_swatch_post": {
        "properties": {
          "checker_patches": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checker Patches"
          },
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_analyze_swatch_foundations_analyze_swatch_post",
        "type": "object"
      },
      "Body_create_foundation_from_photo_foundations_from_photo_post": {
        "properties": {
          "analysis_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Result"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "checker_patches": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checker Patches"
          },
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "shade_code": {
            "default": "",
            "title": "Shade Code",
            "type": "string"
          },
          "shade_name": {
            "title": "Shade Name",
            "type": "string"
          }
        },
        "required": [
          "image",
          "brand",
          "product_name",
          "shade_name"
        ],
        "title": "Body_create_foundation_from_photo_foundations_from_photo_post",
        "type": "object"
      },
...
      "FoundationAnalysisConfidence": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "score",
          "level",
          "notes"
        ],
        "title": "FoundationAnalysisConfidence",
        "type": "object"
      },
      "FoundationAnalysisResult": {
        "properties": {
          "L_value": {
            "title": "L Value",
            "type": "number"
          },
          "a_value": {
            "title": "A Value",
            "type": "number"
          },
          "b_value": {
            "title": "B Value",
            "type": "number"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FoundationAnalysisConfidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "detection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FoundationDetectionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "hex_color": {
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Hex Color",
            "type": "string"
          },
          "undertone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undertone"
          }
        },
        "required": [
          "L_value",
          "a_value",
          "b_value",
          "hex_color"
        ],
        "title": "FoundationAnalysisResult",
        "type": "object"
      },
      "FoundationCreate": {
        "properties": {
          "L_value": {
            "title": "L Value",
            "type": "number"
          },
          "a_value": {
            "title": "A Value",
            "type": "number"
          },
          "b_value": {
            "title": "B Value",
            "type": "number"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "hex_color": {
            "default": "#000000",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Hex Color",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "shade_code": {
            "default": "",
            "title": "Shade Code",
            "type": "string"
          },
          "shade_name": {
            "title": "Shade Name",
            "type": "string"
          },
          "swatch_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swatch Image Url"
          },
          "undertone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undertone"
          }
        },
        "required": [
          "brand",
          "product_name",
          "shade_name",
          "L_value",
          "a_value",
          "b_value"
        ],
        "title": "FoundationCreate",
        "type": "object"
      },
      "FoundationDeleteResult": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "storage_cleanup": {
            "title": "Storage Cleanup",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "storage_cleanup"
        ],
        "title": "FoundationDeleteResult",
        "type": "object"
      },
...
      "FoundationOut": {
        "properties": {
          "L_value": {
            "title": "L Value",
            "type": "number"
          },
          "a_value": {
            "title": "A Value",
            "type": "number"
          },
          "b_value": {
            "title": "B Value",
            "type": "number"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "hex_color": {
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Hex Color",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "shade_code": {
            "title": "Shade Code",
            "type": "string"
          },
          "shade_name": {
            "title": "Shade Name",
            "type": "string"
          },
          "swatch_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swatch Image Url"
          },
          "undertone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undertone"
          }
        },
        "required": [
          "id",
          "brand",
          "product_name",
          "shade_code",
          "shade_name",
          "L_value",
          "a_value",
          "b_value",
          "hex_color",
          "undertone",
          "swatch_image_url"
        ],
        "title": "FoundationOut",
        "type": "object"
      },
      "FoundationUpdate": {
        "properties": {
          "L_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "L Value"
          },
          "a_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "A Value"
          },
          "b_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "B Value"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "hex_color": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hex Color"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "shade_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shade Code"
          },
          "shade_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shade Name"
          },
          "swatch_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swatch Image Url"
          },
          "undertone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undertone"
          }
        },
        "title": "FoundationUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StorageCleanupRetryResult": {
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "completed",
          "pending"
        ],
        "title": "StorageCleanupRetryResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "CIELAB-based skin tone analysis and foundation recommendation",
    "title": "제루미 API",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/foundations": {
      "get": {
        "operationId": "list_foundations_foundations_get",
        "parameters": [
          {
            "in": "query",
            "name": "brand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FoundationOut"
                  },
                  "title": "Response List Foundations Foundations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Foundations",
        "tags": [
          "foundations"
        ]
      },
      "post": {
        "operationId": "create_foundation_foundations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoundationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Foundation",
        "tags": [
          "foundations"
        ]
      }
    },
    "/foundations/analyze-swatch": {
      "post": {
        "description": "Analyze a foundation swatch photo and return extracted color values.\n\nUpload a photo of foundation applied on white paper, optionally with\nColorChecker patches for calibration. Returns LAB values, hex color,\nwithout saving to the database.",
        "operationId": "analyze_swatch_foundations_analyze_swatch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_swatch_foundations_analyze_swatch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Analyze Swatch",
        "tags": [
          "foundations"
        ]
      }
    },
...
    "/foundations/from-photo": {
      "post": {
        "description": "Analyze a foundation swatch photo and save the result to the database.\n\nSame analysis as analyze-swatch, but also creates a Foundation record\nand stores the uploaded image.",
        "operationId": "create_foundation_from_photo_foundations_from_photo_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_foundation_from_photo_foundations_from_photo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Foundation From Photo",
        "tags": [
          "foundations"
        ]
      }
    },
    "/foundations/storage-cleanups/retry": {
      "post": {
        "operationId": "retry_storage_cleanups_foundations_storage_cleanups_retry_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageCleanupRetryResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Retry Storage Cleanups",
        "tags": [
          "foundations"
        ]
      }
    },
    "/foundations/{foundation_id}": {
      "delete": {
        "operationId": "delete_foundation_foundations__foundation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "foundation_id",
            "required": true,
            "schema": {
              "title": "Foundation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationDeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Foundation",
        "tags": [
          "foundations"
        ]
      },
      "get": {
        "operationId": "get_foundation_foundations__foundation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "foundation_id",
            "required": true,
            "schema": {
              "title": "Foundation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Foundation",
        "tags": [
          "foundations"
        ]
      },
      "put": {
        "operationId": "update_foundation_foundations__foundation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "foundation_id",
            "required": true,
            "schema": {
              "title": "Foundation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoundationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Foundation",
        "tags": [
          "foundations"
        ]
      }
    },
...
  }
}